+++ CHANGELOG.md
@@ -19,6 +19,7 @@ TypeScript:
 - Method Observers
   - `connector.subscribe('export', observer: Observer<string>)`
   - `connector.subscribe('import', observer: Observer<DatasetExt[]>)`
+  - `connector.subscribe('*', observer: Observer<any>)`
 - Released versions 1.0.0-alpha.10, 1.0.0-alpha-11
 
 ### Changed